Correction de : - Suppression d'un dossier vide (Issue : 3894) - Ajout d'un nouveau dossier (Issue : 3897)

+++ BananaFlux/branch/V3/srvAjax/manageFolders.php
@@ -5,28 +5,34 @@
     
     if($action == "deleteFolder")
     {
-        $id_folder= htmlentities($_POST['id']);
-        $sqlEmpty = 'DELETE FROM feed_folder WHERE folder_id = "'.mysqli_escape_string($db, $id_folder).'"';
-        $resultEmpty = mysqli_query($db, $sqlEmpty);
+        $id_folder= $_POST['id'];
+        $sqlCheckEmpty = 'SELECT COUNT(*) FROM feed_folder WHERE folder_id = "'.mysqli_escape_string($db, $id_folder).'"';
+        $resultCheckEmpty = mysqli_query($db, $sqlCheckEmpty);
         
-        if(mysqli_affected_rows($db) > 0)
+        $nbFeed = mysql_fetch_row($resultCheckEmpty); 
+        
+        if($nbFeed[0] != 0) 
         {
-            $sql = 'DELETE FROM folder WHERE folder_id = "'.mysqli_escape_string($db, $id_folder).'"';
-            $result = mysqli_query($db, $sql);
+            $sqlEmpty = 'DELETE FROM feed_folder WHERE folder_id = "'.mysqli_escape_string($db, $id_folder).'"';
+            $resultEmpty = mysqli_query($db, $sqlEmpty);
 
-            if (mysqli_affected_rows($db) > 0)
-            {            
-                echo "SUCCESS";
-            }
-            else 
+            if(mysqli_affected_rows($db) < 0)
             {
                 echo $db->error;
-            } 
+            }
         }
-        else
+        
+        $sql = 'DELETE FROM folder WHERE folder_id = "'.mysqli_escape_string($db, $id_folder).'"';
+        $result = mysqli_query($db, $sql);
+
+        if (mysqli_affected_rows($db) > 0)
+        {            
+            echo "SUCCESS";
+        }
+        else 
         {
             echo $db->error;
-        }
+        } 
     }
     else if($action == "editFolder")
     {
@@ -56,10 +62,10 @@
         }
         else
         {
-            header('Location: index.php');
+            return "ERROR_UNLOGGED";
         }        
         
-        $folder_name = htmlentities($_POST['folder_name']);
+        $folder_name = $_POST['folder_name'];
         
         $sql = 'INSERT INTO `folder` (`folder_name`, `user_id`) VALUES ("'.mysqli_escape_string($db, $folder_name).'",'.$user->getID().')';
         $result = mysqli_query($db, $sql);
